✨(api) add support for text queries on topic list requests
We want to create a simple way to search through topics by text.
This can be done through Django's "search" query helper.

The simplest way to implement this is to forget our hierarchy of
topics and just search through plain topics and display them as
a flat list. The hierarchy is then only used when there is no
search params.