🗃(migrations) add a path field to topics for Materialized Path
To be able to display topics according to their depth and in the
expected order, children after their parents, siblings ordered
by name, we need to have their Materialized Path, computed at each
time there is a change in the topics or their hierarchy.

To store it, we need to add a field to the Topic model. We set no
default as it's a mandatory field that should automatically start
filling itself on new Partaj installs.
On existing installs, the default in the migration should be enough
to not crash during the migration.