✨(models) customize the Topic model for Materialized Path computation
We want to make it easy to re-compute Materialized Paths for all topics
whenever we're interacting with topics and we do something which causes
this re-computation to be needed. For this, we can replace the default
model manager for Topic with one that has a builtin method do do that.

We can then override the Topic save and delete methods to call it so
Materialized Paths are automatically maintained whenever there are
relevant updates to topics.