✨(api) update topics API to return children on all list queries
All list queries on topics should return children at the first
level of results, and not only nested through the children
property of their parent.

This makes it easier to consume on the frontend, now that the topics
include their Materialized Path which allows us to know where they
stand in the hierarchy.