✨(frontend) update the topic field in referral form to use autosuggest
As the number of topics in Partaj grows and scales, we cannot rely
on a simple dropdown <select> to enable users to search through them.

We decided to replace the <select> with an autosuggest box that can
be scrolled through for discovery and that will also allow users to
just search for what they need by typing it in.