⚡️(frontend) Improve topic list requests performance
Topic list & search requests were very slow as they packed a lot
of information. The iteration required almost 2000 requests due to
very unoptimized code. We lightened the load and reorganized the
requests to do just 2 requests instead.

A test enforces both this more reasonable number of queries and the
a tolerable execution time.