✨(api) rework serializers to preserve legacy workflow for referral
We reworked the topic serialized to make sure list requests on
topic can be executed in a timely manner.

The topic serialized as used by the ReferralSerialized, on the
other hand, still needs to do more heavy lifting as it is the
passthrough to link it to the unit.

We plan on changing this, linking units directly to referrals, and
re-architecting the app around. Until we do, we can keep a legacy
topic serialized dedicated to this purpose.