💄(frontend) show topic hiearchy in unit topic list
The list of topics for units used to show a flat list. They're now
ordered by path, with the children after their parent.

To reflect the hierarchy, we can use the same trick as in the
autosuggest and pad the left of topic cells depending on the
length of their path, so the hierarchy is represented visually.