✨(api) open endpoints for lists of lite referrals
Requests for lists of referrals are getting too slow on real
environments (staging & prod) as opposed to development environments,
as they are making too many requests, which is costly.

We can solve this issue by making a new type of "lite" referrals
with a minimal number of embedded objects and queries.