♻️(frontend) make the dropdown implementation more generic
We're about to add a new dropdown menu in the sidebar, to allow
users to do actions related to themselves (the first of which
is logging out).

This one does not use the same look for the dropdown opening button.
We decided to refactor our dropdown into a hook to make it more
reusable.