🐛(sidebar) fix a visual glitch on small hight screens
On screens with a small height, the "new referral" button was
moving to the right and hidden behind main content. We want to
prevent wrapping in the sidebar and add scrolling so the user can
move through the navbar-nav.

We also added a shadow on the bottom part as an affordance to let
the user know there is more content to scroll through.