🗃(migrations) add a ManyToMany between referrals & units
Links between referrals and units are currently indirectly managed
through the topic the unit and referral are both linked with, with
two foreignkeys that only allow each referral to have one topic, and
therefore one unit.

It makes it impossible to assign two units to the same referral, and
troublesome to change the one unit assigned as it means changing the
topic to one that is linked with the other unit.

With a ManyToMany, we get much more flexibility in the management of
these relationships. We can still keep the current behavior of Partaj
by automatically creating one of these relationships at the time of
referral creation.