✨(admin) add referral unit assignments to the admin site
As all other models in Partaj, we need to add our new referral unit
assignment model to the admin site.

As it's only used to link referrals & units, we decided to only
expose it as an inline in the referral admin view, which is where
users would look for it.