♻️(models) stop using topic.unit to access referral units
We just added a ManyToMany to link referrals and units directly in
a more flexible way. This means we need to refactor all our use
sites to account for this new relation, which is authoritative as
far as "finding unit(s) for a referral" is concerned.

However, we still need to keep the referral > topic < unit indirect
relation as it's used to determine the initial value to set on
referral.units as the referral is created.