♻️(frontend) update the frontend to stop using topic -> unit
Referrals and units are now linked directly, without going through
the referral topic. This link, while it still exists in the schema,
only serves to point the referral to the right unit upon creation.

This means the frontend needs to be using exclusively referral.units
for any referral related operations.

topic.unit is only useful when doing work on topics themselves.