♻️(api) change back assign/unassign API expectations
We previously modified the API expectations around assignment &
unassignment to require unambiguous requests with the unit for
an assignment and the assignment id for an unassignment.

As it turns out, it is painful for us to rework the frontend to
fulfill these requirements right now. We'd rather keep it this way
and handle that ambiguity later on if it ever becomes necessary.