♻️(frontend) remove the existing <TabGroup /> component
Existing <Tabgroup /> and <Tab /> work with a react state hook,
without taking advantage of react-router.
This is not ideal, and not the way we intend to build tabs in
the referral detail view.

Because of this, we decided to remove <TabGroup /> right now
and move the styles to a CSS class to make it easier to
reuse them, while we implement the referral detail tabs and keep
the existing dashboard index tabs.

We can refactor the dashboard index tabs to use the router later on.