♻️(frontend) update <ReferralDetailContent /> for the new referral view
We're reworking the referral view to get away from our existing
datetime ordered list of activities, towards a more structured view
with tabs where users can find all the relevant information.

As part of this rework, the assignment information for the referral
is to be moved from the referral box to the top of the page. We're
also adding a background to make the referral content more salient in
our UI.