♻️(frontend) create a new component to show referral activities
As part of the referral detail view rework, we're getting rid of the
tracking list with embedded object. We'll still have the tracking view
but only with a list of activities, without the content embedded in.

As a result, we're building the component for this tab, with a reversed
order and a lighter footprint.