♻️(frontend) remove "exact" props on react-router switches
We want <NavLink /> to highlight the active links even when the
user has moved through nested states. In some existing cases this
was prevented by the exact prop/parameter on switches, that
narrowed down the route definition too much.

By removing this argument, we can expand the route definitions and
have them match the active routes.