✨(api) open a list route for referral answers
As part of our ongoing refactor of the referral detail view, we have
a tab with all the answers for the referral, where actions can be
taken and information about their status read by users.

To build that view, we need to add referral answer endpoints as that
will be simpler to handle than the existing referral activities
endpoint.