✨(frontend) add a <ReferralAnswerStatusBadge /> component
As we do with referrals, we want to be able to make the answer
status available at a glance for users. This is a new feature in our
referral detail view rework.

We're implementing as a series of frontend checks on a referral answer
and its validation requests as a first implementation. If we keep it,
we'll probably want to refactor it so it is computed server-side or
even kept in the model itself.