✨(frontend) rework the referral answer validations list
Update the existing referral answer validations list to bring it
up to par with the rest of the Partaj UI in the ongoing general
UI rework.

We split it into its own component to make things clearer as is is
used inside a referral detail route and it holds quite a bit of
complexity towards managing validations requests and performing them.