♻️(frontend) split the answer validation form into its own component
We want to make our structure clearer inside the referral detail nested
router views. To that end, we're separating actual functional components
from the nested routes.

Making the validation form into its own component makes sense in this
effort so we can easily combine it with the referral answer display
to make a view appropriate for users reviewing an answer.