♻️(frontend) rework the referral detail content answer button logic
As we refactor the referral detail view, we decided to implement all
the nesting and tabs logic using react-router. This will enable users
to share links deeper in the app and should help this complex view to
be more reliable overall.

This means we need to rework instances were we used context or state
to manage pseudo-routes to use react-router instead. That is the case
with the answer button in referral detail content.