✅(frontend) add back the <ReferralAnswerValidationsList /> unit tests
<ReferralAnswerValidationlist /> unit tests were included as part of
the <ReferralAnswerDisplay /> tests as that was the only place where
the validations list and the form to add validations appeared.

As we moved things around, it made more sense to us to move the
test to the standalone component instead of testing it as part of yet
another parent component.