✅(frontend) add back the <ReferralAnswerValidationForm /> unit tests
<ReferralAnswerValidationForm /> unit tests were included as part of
the <ReferralAnswerDisplay /> tests as that was the only place where
the validation form appeared.

As we moved things around, it made more sense to us to move the
test to the standalone component instead of testing it as part of yet
another parent component.

We also noticed we were missing a redirect after the user left their
review, and we added it in.