✨(frontend) create a breadcrumbs module for Partaj
The new design for Partaj calls for breadcrumbs to help users figure
out where they are in the app and move around more easily.

We can leverage react-router do build a series of components that
allows us to dynamically add pages to the breadcrumbs as they are
loaded in routers & nested routers, and remove them when they get
off screen.