🗃(migrations) add models for referral messages
We set out to implement requests and answers for more information
on referrals. Instead of differentiating messages by their author,
we chose a more generic implementation where all messages are
instances of the same simple model.

This way, we can implement the current design and update it later
on as needed to fit what users make of those messages. We can use
activity if we want to separate between requests for information and
responses by referral authors.