🔧(env) update test environment base url for easier use in tests
When we want to test scenarios that use the Partaj base url, we
want it to be as simple as possible. In those cases, using a localhost
can create confusion with the dev environment and is not obviously
a test-specific value.

We can use `https://partaj` instead which is very obviously not a
real URL.