⬆️(frontend) update react-query to version 3
We want to start using react query mutations in Partaj. As the API
and types have evolved from version 2 to version 3, we'd rather
use the new version directly rather than build new helpers using the
old APIs.

As types are now included in the react-query package, we do not
need the @types package anymore.