✨(frontend) update referral assignment menu to manage units
Partaj had a referral assignment dropdown menu that allowed unit
organizers to manage member assignments. As we wanted to also
manage unit assignments, we had to rework the button and the dropdown
menu to add tabs so both units and members can be managed from the
same dropdown menu.

We also had to make the dropdown available to non-organizer users
so they can still get information on unit and member assignments
even if they cannot make changes.