🏷️(frontend) add types & factories for referral messages
As we create new models on the backend side, to store & represent
user messages and their related attachments, we need to add matching
types to the frontend so we can make use of them.

We also added factories for them to make it easier for us to write
relevant tests.