🐛(frontend) avoid flickering when a new message is sent in referral
In referral messages, when a new message is sent, we first use a local
representation, then switch to the actual one once the message has
been created on the server.

As a result, we got a flicker in the delay between the end of the
creation of the message, and the moment when the frontend receives
an updated list of messages.

If we only stop showing the queued message when the actual one is
ready to be rendered, we avoid any flicker as there is no render
loop when neither the queued message nor the one from the API are
rendered.