✨(frontend) redirect non-unit users to sent referrals
As they come to Partaj, users that do not belong to a unit were
being redirected to the dashboard view. This makes no sense for them
as they do not have a list of tasks to perform.

Instead, we should redirect them to their list of sent referrals,
which is the one view where they have tasks to perform and information
to receive.