🐛(frontend) fix an issue with referral detail final answer
There was an issue with the referral detail final answer that
resulted in a blank page when coming there from the popin to send
the final answer.

This was caused by the referral not being up to date. Change the
final answer tab to add a query for the referral so we can make
sure to update and get the final answer.