🐛(frontend) remove the "answer referral" button for referral requester
There was an oversight in <ReferralDetailContent /> that did not check
if the current user viewing it was a member of a linked unit, and therefore
someone who could possibly answer the referral, showing the button to
everyone.

Add the relevant check and update our unit tests to make sure we don't
make this mistake again.