✨(frontend) change the assignment dropdown text for requester
When unit members see the assignment dropdown button on a referral
that was not assigned to a user yet, we show them "Not assigned" to
draw their attention to an action they must take.

For the requester, there is no action to take at this time, and this
message can make it look like no one is working on their referral.
We'd rather show them the more neutral "Assignments".