✨(frontend) improve the requester progress bar on referral detail
We recently added a progress bar so referral requesters can tack
the progress of their referral through various steps over time.

We needed to add another additional step in the process. We took
this opportunity to factorize some code into a ProgressBarElement
component and improve accessibility for the progress bar.