💄(chat) always show chat tab text field in referral detail
In the referral detail view's messages tab, the actual text field
to send messages into the chat was hidden below the fold for users
with smaller screens (or who have zoomed in).

This was due to an attempt to make the page "look better" given the
small size of the space allocated to the chat window. We can just
let it flow so it's always shown on the screen.