🐛(api) set attachments on published answers
Referral answer attachments are linked to answers through a ManyToMany
relationship. As published answers are technically separate instances
of ReferralAnswer, linked to the draft by a recursive foreign key,
the link to the attachments needs to be re-created for the published
version of answers.

This was previously not done during publication. We added a test to
prove this and updated the publish transition to create the links.