🐛(frontend) fix referral detail view submenu
The referral detail view submenu shows unwanted additional separators
when the referral can not be closed or cannot have its date changed.
Update the component to hide the separators when the items themselves
are hidden.

Also make sure the link to close the referral is shown on all the
relevant states, including the new PROCESSING and IN_VALIDATION
states.