💄(static) make the logged-off index into a proper landing page
The existing logged-off index was mostly a placeholder we used
because "something" was needed so we just put a bit a text with
a Signup/Login button.

As we work harder on getting more users onto Partaj, we decided
it was time to make a new and improved landing page there to
welcome requesters to Partaj.