(api) send emails for referral closure or urgency level change
When we implemented transitions to enable users to close referrals
or change their urgency levels, we did not create emails to send
them so they can be notified of the changes. This is a problem as
they have no way to see what happens to their referrals other than
checking the referral tracking view.

We can just add those emails on the relevant transitions.