(front) display unit's owners in referral form
We want to help users create their new referrals by making sure
they know who will receive it, and that it will be sent right
away to a real person. To that end, we added the names for the
owners of the target unit as a bit of information directly in the
form.