👷(ci) merge the frontend & backend CI workflows
We initially built two CI definitions, one for the backend and one
for the frontend. This is no longer a good way to organize our CI
jobs as we want to add deployments that depend on both frontend
and backend jobs.

To allow this, we need to have only one CI workflow that runs both
frontend & backend related jobs.