♻️(backend) use the new Referral.users field when appropriate
As we added the new ManyToMany relationship between users and
referrals, we need to update all uses throughout the backend to use
this new relation instead of the existing ForeignKey.

We're also taking this opportunity to get rid of the requester field,
which is not really useful anymore and makes the referral form more
confusing, especially now that it needs to allow users to add more
requesters.