🗃️(migrations) remove legacy referral requester fiel
After we introduced our new ManyToMany Referral.users field, we no
longer need the existing ForeignKey Referral.user nor the legacy
text field Referral.requester.

We're removing the `requester` field right now as it does not have
a lot of value for us, and making the `user` field optional so we
can more easily remove it and revert if necessary.