➕(dependencies) add arrow to backend dependencies
We used the `arrow` package in our actual application code even
though it is not part of our dependencies list. This worked because
it is part of the development dependencies and those happened to be
installed in our container too.

In the Python buildpack, we do not install development dependencies,
and therefore we need to make sure arrow is part of our proper
dependencies.