♻️(frontend) handle stale closure problem in useClickOutside
useClickOutside was creating bugs due to the stale closure problem:
it worked fine when `onClick` never changed during the whole
lifecycle of the hook, but did not take into account updated values
for onClick.

We cannot manage this with a dependency array as we want to avoid
creating a memory leak with a bunch of event listeners. Instead, we
can use a mutable ref to keep an up-to-date onClick reference inside
our closure.