🐛(frontend) fix change urgency level modal
The modal to allow users to change a referral's urgency level
was generating broken changes due to an incorrect value for
the `newUrgencyLevelId` value used in the form that did not
match what was displayed by the select.

We split the component in two so we can set the correct initial
value in our state without creating synchronization problems.
We also removed the existing value from the form.