🍱(icons) add an icon to remove users
As we update the UI to support multiple users on a referral, and
let users add or remove users from referrals, we need an icon
for the button to remove users.

Using a trash bin (!) or even a cross is not suitable for this
purpose, we need a softer icon. This is why we're introducing the
user disconnect icon.