✨(frontend) add a Requesters tab on the referral detail view
As we removed the mention of the requester from the referral
header (it was a simple "Requester: John Doe" string), we still
nedd to be able to display the list of actual requesters on the
referral detail view.

We chose to add a new tab so we can easily show an arbitrarily
large list and manage additions later on.