🐛(deployment) fix ever-expanding static folder
As we built our deployment script, we made it as an additive
process, always adding new files into the repo, never removing
deleted files. This started causing issues as the frontend build
is generating 1600 files per deployment. This is quickly growing
the size of the static folder with unused old files, and
overwhelming our whitenoise which is failing to serve anything.