👷(ci) generate binary l10n files during build
As we migrated Partaj to Scalingo, we added a `compilemessages`
call to the script that is launched by the Procfile, when the
buildpack is starting the app.

For some reason, maybe relating to missing dependencies (gettext),
binary localization files are not generated. We don't control the
buildpack environment. Therefore, we'd rather solve the issue in
an environment we control, which is our own build step in the
deployment CI.