♻️(api) rework referral lite to use get_queryset
As we prepare to add more filters and more kinds of filters on
referral lite list queries, we decided to refactor our referral lite
viewset to move filtering to the get_queryset method.
This will enable us to add more filters in a scalable way and
better manage separation of concerns.

We also removed the separate task-based endpoints and changed tasks
to be regular query parameters. This will enable us to reuse other
filtering facilities while filtering on a given task in the same
request.