✨(forms) add an ArrayField to handle multiple values on a key
As we implement more filters on referral list requests, we need
some of those filters to be able to support multiple values for
a given key as one.

This is where array field comes in: it enables us to perform
validation and deserialization on each of those values, and get a
cleaned list at then end.