✨(api) add state filtering on referral list requests
As we build out our referral list filtering feature, we start by
adding the possibility to filter all lists of referrals (actually
referral lite) by state.

We only test the filter along with unit as testing all possible
combinations of filters (and, later on, sortings) would lead to
an exponential number of tests.