♻️(api) use queryset restriction to manage referral list permission
In some contexts, including the former, simpler referral list
use cases, it is straightforward to manage list permission directly
using the parameters and a bunch of special cases.

As we implement generic, open-ended filtering options, we'd rather
restrict the queryset at the beginning to what the current user can
access, and then simply apply the filters. This is both simpler and
clearer than adding a bunch of permission special cases down into
some filters, and adding more database requests to do so.