♻️(api) enable multi value filtering on referral users & units
As we built Partaj, we needed a way to get lists of referrals for
a given unit, and for a given user. We employed the same tactics,
making an ad-hoc, specific filter for that very purpose.

As we work towards general purpose filtering, we want to be able
to support multiple value filtering on both of those fields. Modify
them to be able to support the same kinds of queries as other filters.