♻️(frontend) rework <ReferralTable /> to load referrals itself
We had a <ReferralTable /> component that only took a list of
referrals and displayed it. As we want to make filters reusable
for several different views, we decided to instead have it load
the list of referrals, and receive a list of params to enable by
default, which it can then complement with what the user enters.