✨(api) add autocomplete filtering to unit lists
As we add a filter for units in referral table, we need a way to
let users pick the unit they want to filter on. As the list can
grow to a decent size, we decided to use an autocompletion field.

We therefore needed to add filtering based on the first characters
of the unit name to the API.