➕(frontend) add react-day-picker to our dependencies
As we build the filters for referral tables, we need to let users
select filters based on due dates. Instead of entering them as
plain text, we decided to use a ready-made component to offer them
a date range picker.

We picked react-day-picker as a full-featured date range picker
that comes with as few dependencies as possible.