✨(frontend) implement filtering in <ReferralTable />
We use referral tables to display list of referrals throughout
Partaj. As we decided to let users filter on referral lists, we
needed to rework it to implement filtering logic.

We used a child component that connects to the ReferralTable
state and uses a form with a date picker and autocompletion fields
to set filters.

We also decided to disable some columns based on context: for
example, in a unit page it makes no sense to let users manipulate
unit filters as that could creation confusion with the purpose
of the page they are viewing.