🐛(api) fix an issue with foreign key based filtering on referrals
As we added a lot of filters on referrals, based on objects related
to referrals, we end up with a bunch of duplicates on some kinds
of queries.

We need to add `.distinct()` to our queryset to make sure we avoid
this duplicate issue. We also modified a unit test to surface the
issue.