🗃️(migrations) add an index field on ReferralUrgency
`is_default` allows us to know which referral urgency to set when
the user did not explicitly set one, and which one to put first
when displaying lists of referral urgencies.

However, in some instances, the ordering of other elements in the
list is mostly undefined (left to defaults), and there is no clear
way to order this list based on an existing field.

We decided to add `index` to replace `is_default` and fill this
gap, with the assumption that the item at `0` index is the default
one.