✨(backend) set up building blocks of Elasticsearch integration
As we integrate Elasticsearch into Partaj, we need to manage
our indices creation. Implement an index manager (really a few
helpers functions to handle looping aliases and indices and
manage errors) so we can regenerate all indices at-will during
development.

It can also serve as an escape hatch to "manage" unresolved bugs
while a proper fix is developed. We added analysis settings as
we will be analyzing full-test for filters and searches.