🐛(api) filter our draft referrals from all but draft requests
As we migrated our referral lists from django querysets to
Elasticsearch queries, we forgot to filter our draft referrals,
which should not appear in queries, except when specifically
requested.

We handled this in two ways:
- this is a rights problem: unit members should not be allowed to
see draft referrals not yet submitted to their unit. In that
spirit, we added a restricting clause on state in the initial state
filter;
- if a user creates a draft that links to a unit they are a member
of (this is an edge case), they would still have permission to see
it there, which could be confusing for them as they would be the only
person on their unit who sees it. To prevent this, we set a default
filter to exclude referrals with state draft whenever referrals are
requester without an explicit state filter.