🐛(search) fix referral object alphabetical sorting
If we naively use sorting on a keyword field with unstructured data,
we end up with an alphabetical ordering that does not match what
users expect: lower VS upper case is ranked differently and
diacritics create yet another sort. We need to normalize the field
to make sorting make sense to users.