✨(search) improve sorting of referrals by state
Sorting referrals by state naively used an alphabetical sorting
for the states. While that can be an acceptable approach, it causes
issues when states are translated: alphabetical order gets broken
because the machine names have nothing to do with the user facing
names for the states.

Since we had to take some steps to improve the ordering, we decided
to implement a logical ordering based on the way referrals go
through the different states.