♻️(search) improve sorting on multi-user columns
When multiple users' names were present in a column used for
alphabetical sorting, sorting was not consistent in how it ordered
users.

We decided to just use the first user in alphabetical order for
the sorting field so we can get more consistent results. We can then
show the relevant users alphabetically sorted inside their columns
on the frontend side so users can see a sorting consistent with
the columns' content.