🐛(api) roll back a change in the behavior of the process filter
As we migrated the referrals "process" filter to our new ES-based
API, we included additional filtering based on the state of the
referral.

As it turns out, this was a mistake as it breaks a feature for some
users where we show them the referrals that have already been
processed. We can just let the client filter by state whenever
appropriate.