🐛(answers) fix an issue with referral answered activities
The referral answered activity needs to display the names of
validators who signed off on the published answer, but they are
not part of the activity object itself.

It previously made an API request on referral answer validation
requests to build that list. However, this creates access rights
issues that are not simple to solve for all use cases, as we do not
want requesters to be able to see the content of validations nor
the attached draft answers.

We decided to embed the list of validators directly on the
referral answer object to sidestep the issue.