✨(indexers) add a helper to set up index field for autocompletion
Often, we want to enable autocompletion to work with other starting
points than the very beginning of the string, typically after a
given delimiter is hit (for example "The example" should autocomplete
on both "The" and "example").

To achieve this, we need to set up substrings from our chosen
delimiter at indexing time so ElasticSearch can set itself up to
perform autocompletion later on.

The slicing helper generates just those substrings that we need to
index in the relevant field of our index.