✨(api) add ES-based list endpoints for units to the API
To improve filtering on lists of referrals (and power up other
features), we decided to create new endpoints to get lists of units
quickly based on various kinds of parameters. To that end,
we're setting up new viewsets that return "lite" units that are
lighter and don't require authorization to view.

As we add autocomplete based filters to the general purpose search
field in <ReferralTable />, we also need to enable autocompletion
for the filter values we're adding in the search field. We can reuse
the Autocomplete mixin we set up for topics.