♻️(frontend) rework referral table filters to store strings only
Referral table filters are stored in-memory for now. This was a
shortcut that allowed us to store full objects instead of IDs so
it's easier to display the filters in human readable formats: the
full objects have enough data to make a human representation.

On the other hand, it prevents us from storing query parameters in
the URL, making navigation quite cumbersome for users.

As we add autosuggested filters, it's becoming overly complicated
to manage a store of whole objects. Instead, we've elected to store
IDs as strings and push the complexity of making human readable
filters in the UI to the component responsible for this UI.