✨(frontend) add autocompletion suggestions to referral table
We have an existing way to add filters by manually picking a
column and then entering a value for that column.

As we added a full-text search field, we decided to use it to
enable autocompletion searches for filters that would match what
is currently entered. This way, we can provide an alternate,
quicker way to let users add filters, that is also more easily
discoverable.